مدیریت تنظیمات و بهبود ساختار پروژه
Working on #87

- Updated `App.xaml.cs` to include a new using directive for `CodeGenerator.Application.Settings`, indicating the introduction of settings management.
- Removed a method for creating a `Property` from a `Field` in `Extensions.cs`, suggesting a refactor or relocation of this functionality.
- Introduced `CodeConfigs.cs`, defining a new class `CodeConfigs` with a `RootNameSpace` property for configuration purposes.
- Modified `CodeFileService.cs` to reference the new settings structure, changing how folder paths are accessed.
- Added a new method `CrateByTable` in `DtoService.cs` to create a `DtoViewModel` from a `Table`, indicating a new approach to data transformation.
- Updated `IServices.cs` to include the new `DtoViewModel` and `CrateByTable` method, reflecting changes in the service interface.
- Added a method for creating a `Property` from a `Field` in `PropertyService.cs`, which was previously removed from `Extensions.cs`.
- Significantly modified `Settings.Extension.cs` to include methods for loading, saving, and configuring settings, indicating a more structured approach to settings management.
- Updated `Settings.cs` to include properties for `Folders` and `CodeConfigs`, suggesting a more organized configuration structure.
- Updated `SelectTableDialog.xaml.cs` and `DtoManagementPage.xaml.cs` to use the new settings structure, aligning UI components with the new configuration management.
- Modified `DtoViewModel.cs` to include the new `CrateByTable` method, changing how DTOs are created and managed.
- Introduced `ProjectLayer.cs`, defining an enumeration for different project layers for better organization.
- Introduced `FolderStructure.cs`, defining a structure for organizing folders within the project, likely tied to the new settings management.
- Updated `IModuleService.cs` and `IPropertyService.cs` to include new methods and using directives, reflecting changes in service interfaces to accommodate the new settings and data structures.
