+++ src/infra/CodeGenerator/Application/Services/IModuleService.cs
@@ -1,0 +1,3 @@
+﻿namespace CodeGenerator.Application.Services;
+
+public interface IModuleService : IViewService<Module>;
