+++ src/infra/CodeGenerator/Application/Settings/ProjectLayer.cs
@@ -1,0 +1,13 @@
+﻿namespace CodeGenerator.Application.Settings;
+
+public enum ProjectLayer
+{
+    None,
+    Page,
+    Component,
+    ViewModel,
+    Controller,
+    Application,
+    ApplicationModel,
+    Repository,
+}
